+++ crates/objc2/src/declare/declare_class_tests.rs
@@ -1,6 +1,7 @@
 #![deny(deprecated, unreachable_code)]
 use core::ptr;
 
+use crate::declare::IvarEncode;
 use crate::rc::{Id, Owned, Shared};
 use crate::runtime::NSObject;
 use crate::{declare_class, extern_methods, sel, ClassType};
@@ -8,10 +9,12 @@ use crate::{declare_class, extern_methods, sel, ClassType};
 // Test that adding the `deprecated` attribute does not mean that warnings
 // when using the method internally are output.
 declare_class!(
+    // Also ensure that empty fields still work
     struct DeclareClassDepreactedMethod {}
 
     unsafe impl ClassType for DeclareClassDepreactedMethod {
         type Super = NSObject;
+        const NAME: &'static str = "DeclareClassDepreactedMethod";
     }
 
     #[deprecated]
@@ -41,6 +44,7 @@ declare_class!(
 
     unsafe impl ClassType for DeclareClassCfg {
         type Super = NSObject;
+        const NAME: &'static str = "DeclareClassCfg";
     }
 
     unsafe impl DeclareClassCfg {
@@ -181,6 +185,7 @@ declare_class!(
 
     unsafe impl ClassType for TestMultipleColonSelector {
         type Super = NSObject;
+        const NAME: &'static str = "TestMultipleColonSelector";
     }
 
     unsafe impl TestMultipleColonSelector {
@@ -252,6 +257,7 @@ declare_class!(
 
     unsafe impl ClassType for DeclareClassAllTheBool {
         type Super = NSObject;
+        const NAME: &'static str = "DeclareClassAllTheBool";
     }
 
     unsafe impl DeclareClassAllTheBool {
@@ -313,6 +319,7 @@ declare_class!(
 
     unsafe impl ClassType for DeclareClassUnreachable {
         type Super = NSObject;
+        const NAME: &'static str = "DeclareClassUnreachable";
     }
 
     // Ensure none of these warn
@@ -352,4 +359,87 @@ declare_class!(
 #[test]
 fn test_unreachable() {
     let _ = DeclareClassUnreachable::class();
+}
+
+#[test]
+#[should_panic = "Failed to add ivar _ivar"]
+fn test_duplicate_ivar() {
+    declare_class!(
+        struct DeclareClassDuplicateIvar {
+            ivar1: IvarEncode<i32, "_ivar">,
+            ivar2: IvarEncode<i32, "_ivar">,
+        }
+
+        mod ivars;
+
+        unsafe impl ClassType for DeclareClassDuplicateIvar {
+            type Super = NSObject;
+            const NAME: &'static str = "DeclareClassDuplicateIvar";
+        }
+    );
+
+    let _ = DeclareClassDuplicateIvar::class();
+}
+
+#[test]
+#[should_panic = "instance variable \"ivar\" already exists on a superclass"]
+fn test_subclass_duplicate_ivar() {
+    declare_class!(
+        struct Cls {
+            ivar_superclass: IvarEncode<i32, "ivar">,
+        }
+
+        mod ivars;
+
+        unsafe impl ClassType for Cls {
+            type Super = NSObject;
+            const NAME: &'static str = "DeclareClassDuplicateIvarSuperclass";
+        }
+    );
+
+    declare_class!(
+        struct SubCls {
+            ivar_subclass: IvarEncode<i32, "ivar">,
+        }
+
+        mod ivars_subclass;
+
+        unsafe impl ClassType for SubCls {
+            type Super = Cls;
+            const NAME: &'static str = "DeclareClassDuplicateIvarSubclass";
+        }
+    );
+
+    extern_methods!(
+        unsafe impl SubCls {
+            #[method_id(new)]
+            fn new() -> Id<Self, Owned>;
+        }
+    );
+
+    let _ = SubCls::class();
+
+    // The rest is just to show what would go wrong if it didn't panic
+
+    assert_eq!(Cls::class().instance_size(), 16);
+    assert_eq!(SubCls::class().instance_size(), 16);
+
+    let mut obj = SubCls::new();
+
+    // Zero-initialized
+    assert_eq!(*obj.ivar_superclass, 0);
+    assert_eq!(*obj.ivar_subclass, 0);
+
+    *obj.ivar_subclass = 2;
+
+    assert_eq!(*obj.ivar_superclass, 2);
+    assert_eq!(*obj.ivar_subclass, 2);
+
+    *obj.ivar_superclass = 3;
+
+    assert_eq!(*obj.ivar_superclass, 3);
+    assert_eq!(*obj.ivar_subclass, 3);
+
+    let ivar_dynamically = unsafe { obj.ivar::<i32>("ivar") };
+    assert_eq!(*ivar_dynamically, 3);
 }