+++ crates/objc2/src/declare/ivar.rs
@@ -1,10 +1,10 @@
 use core::fmt;
 use core::marker::PhantomData;
-use core::mem::MaybeUninit;
+use core::mem::{self, MaybeUninit};
 use core::ops::{Deref, DerefMut};
 use core::ptr::{self, NonNull};
 
-use crate::encode::{EncodeConvert, Encoding};
+use crate::encode::Encode;
 use crate::runtime::{ivar_offset, Object};
 
 pub(crate) mod private {
@@ -14,9 +14,9 @@ pub(crate) mod private {
 /// Types that may be used in ivars.
 ///
 /// This may be either:
-/// - [`bool`].
+/// - [`IvarBool`][super::IvarBool].
 /// - [`IvarDrop<T>`][super::IvarDrop].
-/// - Something that implements [`Encode`][crate::Encode].
+/// - [`IvarEncode<T>`][super::IvarEncode].
 ///
 /// This is a sealed trait, and should not need to be implemented. Open an
 /// issue if you know a use-case where this restrition should be lifted!
@@ -25,60 +25,34 @@ pub(crate) mod private {
 /// # Safety
 ///
 /// You cannot rely on any safety guarantees from this.
-pub unsafe trait InnerIvarType: private::Sealed {
-    #[doc(hidden)]
-    const __ENCODING: Encoding;
-
-    // SAFETY: It must be safe to transmute from `__Inner` to `Output`.
-    #[doc(hidden)]
-    type __Inner;
-
+//
+// The type must have the same memory layout as the output type.
+//
+// Additionally, the type must be safe to drop even if zero-initialized.
+//
+// Note that while I couldn't find a reference on whether ivars are
+// zero-initialized or not, it has been true since the Objective-C version
+// shipped with Mac OS X 10.0 [link] and is generally what one would expect
+// coming from C. So I think it's a valid assumption to make!
+//
+// [link]: https://github.com/apple-oss-distributions/objc4/blob/objc4-208/runtime/objc-class.m#L367
+pub unsafe trait InnerIvarType: private::Sealed + Encode {
     /// The type that an `Ivar` containing this will dereference to.
     ///
     /// E.g. `Ivar<IvarDrop<Box<u8>>>` will deref to `Box<u8>`.
     type Output;
 
-    // SAFETY: The __Inner type must be safe to drop even if zero-initialized.
+    /// # Safety
+    ///
+    /// The instance variable must have been initialized.
     #[doc(hidden)]
-    const __MAY_DROP: bool;
+    unsafe fn __deref(&self) -> &Self::Output;
 
+    /// # Safety
+    ///
+    /// The instance variable must have been initialized.
     #[doc(hidden)]
-    unsafe fn __to_ref(inner: &Self::__Inner) -> &Self::Output;
-
-    #[doc(hidden)]
-    unsafe fn __to_mut(inner: &mut Self::__Inner) -> &mut Self::Output;
-
-    #[doc(hidden)]
-    fn __to_ptr(inner: NonNull<Self::__Inner>) -> NonNull<Self::Output>;
-}
-
-impl<T: EncodeConvert> private::Sealed for T {}
-unsafe impl<T: EncodeConvert> InnerIvarType for T {
-    const __ENCODING: Encoding = <Self as EncodeConvert>::__ENCODING;
-    type __Inner = Self;
-    type Output = Self;
-    // Note: We explicitly tell `Ivar` that it shouldn't do anything to drop,
-    // since if the object was deallocated before an `init` method was called,
-    // the ivar would not have been initialized properly!
-    //
-    // For example in the case of `NonNull<u8>`, it would be zero-initialized
-    // which is an invalid state for that.
-    const __MAY_DROP: bool = false;
-
-    #[inline]
-    unsafe fn __to_ref(inner: &Self::__Inner) -> &Self::Output {
-        inner
-    }
-
-    #[inline]
-    unsafe fn __to_mut(inner: &mut Self::__Inner) -> &mut Self::Output {
-        inner
-    }
-
-    #[inline]
-    fn __to_ptr(inner: NonNull<Self::__Inner>) -> NonNull<Self::Output> {
-        inner
-    }
+    unsafe fn __deref_mut(&mut self) -> &mut Self::Output;
 }
 
 /// Helper trait for defining instance variables.
@@ -99,13 +73,13 @@ unsafe impl<T: EncodeConvert> InnerIvarType for T {
 /// Create an instance variable `myCustomIvar` with type `i32`.
 ///
 /// ```
-/// use objc2::declare::IvarType;
+/// use objc2::declare::{IvarEncode, IvarType};
 ///
 /// // Helper type
 /// struct MyCustomIvar;
 ///
 /// unsafe impl IvarType for MyCustomIvar {
-///     type Type = i32;
+///     type Type = IvarEncode<i32>;
 ///     const NAME: &'static str = "myCustomIvar";
 /// }
 ///
@@ -120,7 +94,7 @@ pub unsafe trait IvarType {
     #[doc(hidden)]
     unsafe fn __offset(ptr: NonNull<Object>) -> isize {
         let obj = unsafe { ptr.as_ref() };
-        ivar_offset(obj.class(), Self::NAME, &Self::Type::__ENCODING)
+        ivar_offset(obj.class(), Self::NAME, &Self::Type::ENCODING)
     }
 }
 
@@ -166,13 +140,13 @@ pub unsafe trait IvarType {
 /// # Examples
 ///
 /// ```
-/// use objc2::declare::{Ivar, IvarType};
+/// use objc2::declare::{Ivar, IvarEncode, IvarType};
 /// use objc2::runtime::Object;
 ///
 /// // Declare ivar with given type and name
 /// struct MyCustomIvar;
 /// unsafe impl IvarType for MyCustomIvar {
-///     type Type = i32;
+///     type Type = IvarEncode<i32>;
 ///     const NAME: &'static str = "myCustomIvar";
 /// }
 ///
@@ -203,15 +177,13 @@ pub struct Ivar<T: IvarType> {
     /// Make this type allowed in `repr(C)`
     inner: [u8; 0],
     /// For proper variance and auto traits
-    item: PhantomData<<T::Type as InnerIvarType>::Output>,
+    item: PhantomData<T::Type>,
 }
 
 impl<T: IvarType> Drop for Ivar<T> {
     #[inline]
     fn drop(&mut self) {
-        if <T::Type as InnerIvarType>::__MAY_DROP {
-            // SAFETY: We drop the inner type, which is guaranteed by
-            // `__MAY_DROP` to always be safe to drop.
+        if mem::needs_drop::<T::Type>() {
             unsafe { ptr::drop_in_place(self.as_inner_mut_ptr().as_ptr()) }
         }
     }
@@ -225,11 +197,11 @@ impl<T: IvarType> Ivar<T> {
     ///
     /// This is similar to [`MaybeUninit::as_ptr`], see that for usage
     /// instructions.
-    pub fn as_ptr(this: &Self) -> *const <T::Type as InnerIvarType>::Output {
-        T::Type::__to_ptr(this.as_inner_ptr()).as_ptr()
+    pub fn as_ptr(this: &Self) -> *const <Self as Deref>::Target {
+        this.as_inner_ptr().as_ptr().cast()
     }
 
-    fn as_inner_ptr(&self) -> NonNull<<T::Type as InnerIvarType>::__Inner> {
+    fn as_inner_ptr(&self) -> NonNull<T::Type> {
         let ptr: NonNull<Object> = NonNull::from(self).cast();
 
         // SAFETY: The user ensures that this is placed in a struct that can
@@ -245,7 +217,7 @@ impl<T: IvarType> Ivar<T> {
         // so that is fine.
         let offset = unsafe { T::__offset(ptr) };
         // SAFETY: The offset is valid
-        unsafe { Object::ivar_at_offset::<<T::Type as InnerIvarType>::__Inner>(ptr, offset) }
+        unsafe { Object::ivar_at_offset::<T::Type>(ptr, offset) }
     }
 
     /// Get a mutable pointer to the instance variable.
@@ -258,17 +230,17 @@ impl<T: IvarType> Ivar<T> {
     ///
     /// This is similar to [`MaybeUninit::as_mut_ptr`], see that for usage
     /// instructions.
-    pub fn as_mut_ptr(this: &mut Self) -> *mut <T::Type as InnerIvarType>::Output {
-        T::Type::__to_ptr(this.as_inner_mut_ptr()).as_ptr()
+    pub fn as_mut_ptr(this: &mut Self) -> *mut <Self as Deref>::Target {
+        this.as_inner_mut_ptr().as_ptr().cast()
     }
 
-    fn as_inner_mut_ptr(&mut self) -> NonNull<<T::Type as InnerIvarType>::__Inner> {
+    fn as_inner_mut_ptr(&mut self) -> NonNull<T::Type> {
         let ptr: NonNull<Object> = NonNull::from(self).cast();
 
         // SAFETY: Same as `as_inner_ptr`
         let offset = unsafe { T::__offset(ptr) };
         // SAFETY: The offset is valid
-        unsafe { Object::ivar_at_offset::<<T::Type as InnerIvarType>::__Inner>(ptr, offset) }
+        unsafe { Object::ivar_at_offset::<T::Type>(ptr, offset) }
     }
 
     /// Sets the value of the instance variable.
@@ -278,12 +250,9 @@ impl<T: IvarType> Ivar<T> {
     ///
     /// This is similar to [`MaybeUninit::write`], see that for usage
     /// instructions.
-    pub fn write(
-        this: &mut Self,
-        val: <T::Type as InnerIvarType>::Output,
-    ) -> &mut <T::Type as InnerIvarType>::Output {
-        let ptr: *mut MaybeUninit<<T::Type as InnerIvarType>::Output> =
-            Self::as_mut_ptr(this).cast();
+    pub fn write(this: &mut Self, val: <Self as Deref>::Target) -> &mut <Self as Deref>::Target {
+        let ptr: *mut <T::Type as InnerIvarType>::Output = Self::as_mut_ptr(this);
+        let ptr: *mut MaybeUninit<<T::Type as InnerIvarType>::Output> = ptr.cast();
         let ivar = unsafe { ptr.as_mut().unwrap_unchecked() };
         ivar.write(val)
     }
@@ -299,7 +268,7 @@ impl<T: IvarType> Deref for Ivar<T> {
         //
         // Since all accesses to a particular ivar only goes through one
         // `Ivar`, if we have `&Ivar` we know that `&T` is safe.
-        unsafe { T::Type::__to_ref(self.as_inner_ptr().as_ref()) }
+        unsafe { self.as_inner_ptr().as_ref().__deref() }
     }
 }
 
@@ -330,7 +299,7 @@ impl<T: IvarType> DerefMut for Ivar<T> {
         //
         // And using `mut` would create aliasing mutable reference to the
         // object.
-        unsafe { T::Type::__to_mut(self.as_inner_mut_ptr().as_mut()) }
+        unsafe { self.as_inner_mut_ptr().as_mut().__deref_mut() }
     }
 }
 
@@ -348,6 +317,7 @@ mod tests {
     use std::sync::atomic::{AtomicBool, Ordering};
 
     use super::*;
+    use crate::declare::{IvarBool, IvarEncode};
     use crate::rc::{Id, Owned};
     use crate::runtime::NSObject;
     use crate::{declare_class, msg_send, msg_send_id, test_utils, ClassType, MessageReceiver};
@@ -355,7 +325,7 @@ mod tests {
     struct TestIvar;
 
     unsafe impl IvarType for TestIvar {
-        type Type = u32;
+        type Type = IvarEncode<u32>;
         const NAME: &'static str = "_foo";
     }
 
@@ -396,24 +366,26 @@ mod tests {
         declare_class!(
             #[derive(Debug, PartialEq, Eq)]
             struct CustomDrop {
-                ivar: u8,
+                ivar: IvarEncode<u8, "_ivar">,
+                ivar_bool: IvarBool<"_ivar_bool">,
             }
+
+            mod customdrop;
 
             unsafe impl ClassType for CustomDrop {
                 type Super = NSObject;
-            }
-
-            unsafe impl CustomDrop {
-                #[method(dealloc)]
-                fn dealloc(&mut self) {
-                    HAS_RUN_DEALLOC.store(true, Ordering::SeqCst);
-                    unsafe { msg_send![super(self), dealloc] }
-                }
+                const NAME: &'static str = "CustomDrop";
             }
         );
 
+        impl Drop for CustomDrop {
+            fn drop(&mut self) {
+                HAS_RUN_DEALLOC.store(true, Ordering::Relaxed);
+            }
+        }
+
         let _: Id<CustomDrop, Owned> = unsafe { msg_send_id![CustomDrop::class(), new] };
 
-        assert!(HAS_RUN_DEALLOC.load(Ordering::SeqCst));
+        assert!(HAS_RUN_DEALLOC.load(Ordering::Relaxed));
     }
 }