+++ crates/objc2/src/rc/test_object.rs
@@ -57,6 +57,7 @@ declare_class!(
 
     unsafe impl ClassType for __RcTestObject {
         type Super = NSObject;
+        const NAME: &'static str = "__RcTestObject";
     }
 
     unsafe impl __RcTestObject {
@@ -122,12 +123,6 @@ declare_class!(
         fn autorelease(&self) -> *mut Self {
             TEST_DATA.with(|data| data.borrow_mut().autorelease += 1);
             unsafe { msg_send![super(self), autorelease] }
-        }
-
-        #[method(dealloc)]
-        unsafe fn dealloc(&mut self) {
-            TEST_DATA.with(|data| data.borrow_mut().dealloc += 1);
-            unsafe { msg_send![super(self), dealloc] }
         }
 
         #[method(_tryRetain)]
@@ -272,6 +267,12 @@ declare_class!(
     }
 );
 
+impl Drop for __RcTestObject {
+    fn drop(&mut self) {
+        TEST_DATA.with(|data| data.borrow_mut().dealloc += 1);
+    }
+}
+
 unsafe impl Send for __RcTestObject {}
 unsafe impl Sync for __RcTestObject {}
 
@@ -290,6 +291,7 @@ declare_class!(
     unsafe impl ClassType for RcTestObjectSubclass {
         #[inherits(NSObject)]
         type Super = __RcTestObject;
+        const NAME: &'static str = "RcTestObjectSubclass";
     }
 );