+++ crates/objc2/tests/declare_class_self.rs
@@ -25,6 +25,7 @@ declare_class!(
 
     unsafe impl ClassType for MyTestObject {
         type Super = NSObject;
+        const NAME: &'static str = "MyTestObject";
     }
 
     unsafe impl MyTestObject {