+++ crates/icrate/CHANGELOG.md
@@ -87,6 +87,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   `Shared`.
 * Removed a few instances of `TodoProtocols`.
 * Fixed type-encoding of a few `struct`s.
+* Fixed `NSProxy` trait methods.
 
 
 ## icrate 0.0.1 - 2022-12-24